Task 2

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,7 @@
 # manifest that kills a process named killmenow
-exec {'killmenow':
-	command => 'pkill killmenow',
-	path    => '/usr/bin:/usr/sbin:/bin'
+exec { 'killmenow':
+  command => 'pkill killmenow',
+  path    => '/usr/bin:/usr/sbin:/bin',
+  onlyif  => 'pgrep killmenow',
 }
+
